Warning: rename vimfiles to .vim
This change unifies console and gvim so they can look into the same
folder. This removes the need to duplicate the vimfiles folder to get
access to the packages and other syntax and syntax highlighting.

If random things appear to be breaking, this is a good change to look at
as the culprit for a while.